+++ R/color_scale.R
@@ -39,9 +39,11 @@ color_scale <- function(param, zlim, na.value = "transparent") {
                           "BACKGROUND",
                           "WEATHER",
                           "BIOLOGY",
-                          "CELL")) {
+                          "CELL",
+                          "HGHT",
+                          "HEIGHT")) {
     colorscale <-
-      viridis::scale_colour_viridis(na.value = na.value, name = param)
+      viridis::scale_colour_viridis(na.value = na.value, name = param, limits = zlim)
   } else {
     colorscale <- ggplot2::scale_colour_gradientn(
       colours = colors_dbz,
@@ -62,8 +64,10 @@ color_scale_fill <- function(param, zlim, na.value = "transparent") {
                           "BACKGROUND",
                           "WEATHER",
                           "BIOLOGY",
-                          "CELL")) {
-    colorscale <- viridis::scale_fill_viridis(na.value = na.value, name = param)
+                          "CELL",
+                          "HGHT",
+                          "HEIGHT")) {
+    colorscale <- viridis::scale_fill_viridis(na.value = na.value, name = param, limits = zlim)
   } else {
     colorscale <-
       color_palette_to_scale_fill(param,
@@ -168,6 +172,9 @@ get_zlim <- function(param, zlim) {
   if (param == "PHIDP") {
     return(c(-200, 200))
   }
+  if (param %in% c("HGHT","HEIGHT")) {
+    return(c(0, 1000))
+  }
   if (param %in% c("vid", "VID")) {
     return(c(0, 200))
   }
@@ -179,6 +186,9 @@ get_zlim <- function(param, zlim) {
   }
   if (param %in% c("eta_sum", "eta_sum_expected")) {
     return(c(0, 2000))
+  }
+  if (param %in% c("eta_sum_to_VIR")) {
+    return(c(0, 10))
   }
   if (param == "overlap") {
     return(c(0, 1))