+++ NEWS.md
@@ -1,10 +1,33 @@
 # bioRad 0.11.0.9000 (development version)
 
+## New features
+
+* `calculate_vp()` now supports profile estimation with height bins defined relative to ground level and antenna level (#686).
+
+* `integrate_to_ppi()` now supports range correction with profiles referenced to ground level, accounting for local topography (#747, #748).
+
+* New function `add_param()` for mapping raster data onto `scan` and `pvol` object (#767).
+
+* New output field `eta_sum_to_VIR` in `integrate_to_ppi()` (#748).
+
+* New additional arguments `alt_min` and `filter_all_heights` in `filter_precip()` for precip filtering at high altitudes only (#754, #755).
+
+* `beam_width()` and related functions have a new argument `path` for outputting either the one-way or
+two-way beam width of the antenna pattern. The default is changed to two-way (formerly one-way) (#744).
+
+* Expose `eta_max parameter`  as a user option in `calculate_vp()`, to set maximum bird reflectivity (maps to option `etaMax` in `vol2birdR::vol2bird_config()`) (#763)
+
+* Expose `max_nyquist_dealias` parameter to set maximum Nyquist velocity above which to suppress dealiasing (maps to option `maxNyquistDealias` in `vol2birdR::vol2bird_config()`) (#684).
+
+* `raster` argument of `integrate_to_ppi()`, `project_as_ppi()` and `scan_convert` now accepts terra:SpatRaster (8f8458a) 
+
+## Bugfixes
+
+* Bugfix parsing `azim_min`, `azim_max` and `elev_max` arguments for pvol objects (#762).
+
 * Changed lower bound of allowed values for `nyquist_min` argument of `calculate_vp()` to zero (#761).
 
 * Bugfix for `as.vpts()` fixing uninformative error message in the case of NA values in the `source_file` column of a vpts data.frame (#759).
-
-* New additional arguments `alt_min` and `filter_all_heights` in `filter_precip()` for precip filtering at high altitudes only (#755).
 
 * Bugfix for `filter_precip()` not being applied to `vp` objects (#755).
 
@@ -14,8 +37,6 @@
 
 * `eta_to_dbz()` now accepts NA or NaN input reflectivity values (#741).
 
-* `beam_width()` and related functions have a new argument `path` for outputting either the one-way or
-two-way beam width of the antenna pattern. The default is changed to two-way (formerly one-way) (#744).
 
 # bioRad 0.11.0